Traversy- 2:47

+++ traversy/application/views.py
@@ -1,8 +1,13 @@
 from django.shortcuts import render, redirect
-# from django.http import HttpResponse
+from django.http import HttpResponse
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+
 import pprint
 from django.forms.models import model_to_dict
 
@@ -27,6 +32,7 @@ def room(request, id):
     return render(request, 'room.html', { 'data': room })
 
 
+@login_required(login_url='login')
 def create(request):
     form = RoomForm()
 
@@ -40,6 +46,7 @@ def create(request):
     return render(request, 'form_room.html', context)
 
 
+@login_required(login_url='login')
 def update(request, id):
     room = Room.objects.get(id=id)
     form = RoomForm(instance=room)
@@ -54,9 +61,13 @@ def update(request, id):
     return render(request, 'form_room.html', context)
 
 
+@login_required(login_url='login')
 def delete(request, id):
     room = Room.objects.get(id=id)
 
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
@@ -64,3 +75,38 @@ def delete(request, id):
     context = { 'data': id }
     return render(request, 'delete.html', context)
 
+
+def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        userName = request.POST.get('username')
+        email = request.POST.get('email')
+        password = request.POST.get('password')
+
+        try:
+            #user = User.objects.get(email=email)
+            user = User.objects.get(username=userName)
+        except:
+            messages.error(request, 'User does not exist')
+
+        userAuthenticated = authenticate(request, username=userName, password=password)
+        
+        print(userAuthenticated, user.email)
+
+        if userAuthenticated is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR password are wrong')
+
+    context = { 'page': page }
+    return render(request, 'login.html', context )
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
